// frontend2/RentingServiceFrontend/src/app/interfaces/createForSalePostDto.ts
export interface createForSalePostDto {
  Title: string;
  Description: string;
  MainCategory: string;
  SquareFootage: number;
  Price: number;
  PicturesPath: string;
  Lat: string;
  Lng: string;
  BuildingNumber: string;
  Street: string;
  District: string;
  City: string;
  Country: string;
}